+++ src/test/java/dk/tbsalling/aismessages/ais/messages/InterrogationTest.java
@@ -23,11 +23,11 @@ public class InterrogationTest {
         // Assert
         assertEquals(AISMessageType.Interrogation, aisMessage.getMessageType());
         Interrogation message = (Interrogation) aisMessage;
-        assertEquals(Integer.valueOf(3), message.getRepeatIndicator());
+        assertEquals(3, message.getRepeatIndicator());
         assertEquals(MMSI.valueOf(3669708), message.getSourceMmsi());
         assertEquals(MMSI.valueOf(366969740), message.getInterrogatedMmsi1());
-        assertEquals((Integer) 0, message.getType1_1());
-        assertEquals((Integer) 0, message.getOffset1_1());
+        assertEquals(0, message.getType1_1());
+        assertEquals(0, message.getOffset1_1());
         assertEquals(-1, message.getType1_2());
         assertEquals(-1, message.getOffset1_2());
         assertNull(message.getInterrogatedMmsi2());
